# api.yaml
openapi: 3.0.3

info:
  title: BUPT Software Engineering Homework API
  description: BUPT Software Engineering Homework API
  version: 1.0.0

servers:
  - url: "http://localhost:11451/api"
    description: Local server

security:
  - sessionAuth: []

paths:
  /login:
    post:
      tags:
        - Authorization
      summary: Login as admin
      parameters:
        - in: query
          name: no-csrf # If the server doesn't implement CSRF token, add this parameter to disable CSRF token
          schema:
            type: boolean
          description: Disable CSRF token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin
      responses:
        "200":
          description: Login successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=abcde12345; Path=/
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: admin
                  csrf_token: # Optional: anti-CSRF token
                    type: string
                    example: abcde12345
        "401":
          description: Login failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"
  /logout:
    post:
      tags:
        - Authorization
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      summary: Logout
      responses:
        "204":
          description: Logout successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session=; Path=/
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"

  /record:
    post:
      tags:
        - Data Collect
      summary: Record data
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/record_entry"
      responses:
        "200":
          description: Record successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  unique_id:
                    type: string
                    example: 1145141919810abc # random Unique ID, 16 characters, same as the input
        "403":
          description: Signature not match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"

  /admin/device:
    put:
      tags:
        - Administrate
      summary: Add device
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                room:
                  type: string
                  example: 2-233
                public_key:
                  type: string # RSA 4096
      responses:
        "200":
          description: Add device successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    type: string
                    example: 2-233 # same as the input
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"
    delete:
      tags:
        - Administrate
      summary: Remove device
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                room:
                  type: string
                  example: 2-233
      responses:
        "200":
          description: Remove device successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    type: string
                    example: 2-233 # same as the input
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"
  /admin/devices:
    get:
      tags:
        - Administrate
      summary: Get all available devices
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      responses:
        "200":
          description: Get devices successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: 2-233
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"

  /status/{room_id}:
    get:
      tags:
        - Data Query
      summary: Get status of a device
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: room_id
          schema:
            type: string
            example: 2-233
          required: true
          description: Room ID
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      responses:
        "200":
          description: Get status successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    type: string
                    example: 2-233
                  temperature:
                    type: integer
                    example: 26
                  wind_speed:
                    type: integer
                    example: 3
                  mode:
                    type: string
                    example: cold
                  sweep:
                    type: boolean
                    example: true
                  is_on:
                    type: boolean
                    example: true
                  last_update:
                    type: string
                    example: 2023-09-18T11:45:14+08:00 # ISO 8601
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"
  /status:
    get:
      tags:
        - Data Query
      summary: Get status of all devices
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      responses:
        "200":
          description: Get status successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    room:
                      type: string
                      example: 2-233
                    is_on:
                      type: boolean
                      example: true
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"

  /room/check_in:
    post:
      tags:
        - Room
      summary: Check-in a room
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                room:
                  type: string
                  example: 2-233
      responses:
        "200":
          description: Check-in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    type: string
                    example: 2-233 # same as the input
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"
  /room/check_out:
    post:
      tags:
        - Room
      summary: Check-out a room
      security:
        - sessionAuth: []
      parameters:
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345 # Optional: anti-CSRF token
          description: CSRF token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                room:
                  type: string
                  example: 2-233
      responses:
        "200":
          description: Check-out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    type: string
                    example: 2-233 # same as the input
                  report:
                    $ref: "#/components/schemas/report_data"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error_response"

  /statistics:
    get:
      tags: 
        - Data Statistics
      summary: Get cost and income statistics within a period
      security:
        - sessionAuth: []
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
            example: "2023-01-01"
          required: true
          description: Start date
        - in: query  
          name: end_date
          schema:
            type: string
            format: date
            example: "2023-12-31" 
          required: true
          description: End date
        - in: header
          name: X-CSRF-Token
          schema:
            type: string
            example: abcde12345
          description: CSRF token  
      responses:
        "200":
          description: Get statistics successfully
          content:
            application/json:
              schema:
                type: object
                properties: 
                  total_cost:
                    type: integer
                    example: 1000000
                  total_income: 
                    type: integer
                    example: 1100000
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/error_response"

components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie # Optional: use CSRF token, header = X-CSRF-Token
      name: session
  schemas:
    error_response:
      type: object
      properties:
        error_code:
          type: integer
          example: 100
    record_entry:
      type: object
      properties:
        room:
          type: string
          example: 2-233
        operation:
          type: string
          example: start_26_3
          # start_<temperature_number>_<wind_speed_number>_<mode>_<sweep>, stop,
          # temperature_<number>, wind_speed_<number>, mode_<mode>, sweep_<on/off>
        time:
          type: string
          example: 2023-09-18T11:45:14+08:00 # ISO 8601
        unique_id:
          type: string
          example: 1145141919810abc # random Unique ID, 16 characters
        signature:
          type: string
          example: RTMiQudovtn1Vz8FTVr7LBRocgd94twScB8u1AvaYAsS7kskXmHFs1KvHF9vsslS0uV3gfdyJFrH1HBj/kKzf2Z/y3XPJFMs90uqqN1JHDOwvTytxC2Dl8rL53LHD/fcbVsJa/I28AcxYXBiz79t/WYLhEIUBMrDjCXXYIZLvaQgEonsA4/Zb5kT5R1TRRgrAu9jHbqQU1dwPp0T6tUSZBVboJ8wSokRuUDCbWAQSW34pu3r+DVSNsWpZ5rN4xah/xo2J6A4UhNwvbqwoefbjfwO6xwjLyzN5VVZFzPUbgAF8ttn7uWfXJFhjhkRmPPCd5g3F5p74MCCdwdmCcA0MXjAmajtCj7KBw8x4WZ6Emb8t7z8uFqHdzh9ZLYFk4DBlkkTdUlHgwG40om29XbqCQgsUipXB9+v+c6q00oNukRQYthCARaaSIFwCxGPeAuAl+YJy5Y5fHBLnWYOF4faEFiueM0pK+NjRqmKn9LwYrOsSmENwNm959sK9Kb2qljCWIurTjjHvcxEriMMfJsZARXD7a1Ms7cGV0ABHQIhvIdOmtjKE+qWvF6kbbeZ5+nikNP+tJufnlFE3vU03oSQGPPnYdUz6juvQIKHXb5xe8j1VrnsYyZTD/5HhblhydTGZxzYUdmFjk/rs2BrMscEZPAzgBgANJw/k2FD/70CxLg=
          # SHA256withRSA, RSA 4096, sign text = room + operation + time + unique_id
    report_data:
      type: object
      properties:
        total_cost:
          type: integer
          example: 114514
        total_time:
          type: integer
          example: 1919810
        details:
          type: array
          items:
              type: object
              properties:
                start_time:
                  type: string
                  example: 2023-09-18T11:45:14+08:00 # ISO 8601
                end_time:
                  type: string
                  example: 2023-09-19T11:45:14+08:00 # ISO 8601
                temperature:
                  type: integer
                  example: 26
                wind_speed:
                  type: integer
                  example: 3
                mode:
                  type: string
                  example: cold
                sweep:
                  type: boolean
                  example: true
                duration:
                  type: integer
                  example: 86400 # seconds
                cost:
                  type: integer
                  example: 114514